different templates extracted as adapters
